refactor: extract post entry into a separate component
