+++ ui/apps/akasha/src/components/post-page/post-page.tsx
@@ -3,36 +3,24 @@ import { useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 
 import DS from '@akashaorg/design-system';
-import {
-  RootComponentProps,
-  EntityTypes,
-  ModalNavigationOptions,
-  AnalyticsCategories,
-} from '@akashaorg/typings/ui';
-
-import { ILocale } from '@akashaorg/design-system/lib/utils/time';
+import { RootComponentProps, EntityTypes, ModalNavigationOptions } from '@akashaorg/typings/ui';
 
 import {
   LoginState,
-  useIsFollowingMultiple,
-  useFollow,
-  useUnfollow,
   useInfiniteComments,
   usePost,
   mapEntry,
   useGetProfile,
-  useEntryNavigation,
 } from '@akashaorg/ui-awf-hooks';
 
 import FeedWidget from '@akashaorg/ui-lib-feed/lib/components/App';
-import routes, { POST } from '../../routes';
 import { useAnalytics } from '@akashaorg/ui-awf-hooks';
 import { Extension } from '../extension';
+import { PostEntry } from './post-entry';
 
 const {
   Box,
   BasicCardBox,
-  EntryBox,
   Helmet,
   EditorPlaceholder,
   EntryCardHidden,
@@ -61,6 +49,7 @@ const PostPage: React.FC<IPostPageProps & RootComponentProps> = props => {
     loggedUser: loginState?.pubKey,
     enabler: loginState?.fromCache,
   });
+
   const entryData = React.useMemo(() => {
     if (postReq.data) {
       return mapEntry(postReq.data);
@@ -85,59 +74,13 @@ const PostPage: React.FC<IPostPageProps & RootComponentProps> = props => {
     return [];
   }, [reqComments.data]);
 
-  const locale = (props.plugins['@akashaorg/app-translation']?.translation?.i18n?.languages?.[0] ||
-    'en') as ILocale;
-
   const profileDataReq = useGetProfile(loginState?.pubKey);
   const loggedProfileData = profileDataReq.data;
-
-  const isFollowingMultipleReq = useIsFollowingMultiple(loginState?.pubKey, [
-    entryData?.author?.pubKey,
-  ]);
-  const followedProfiles = isFollowingMultipleReq.data;
-  const followReq = useFollow();
-  const unfollowReq = useUnfollow();
-
-  const handleFollow = () => {
-    if (entryData?.author.pubKey) {
-      followReq.mutate(entryData?.author.pubKey);
-    }
-  };
-
-  const handleUnfollow = () => {
-    if (entryData.author.pubKey) {
-      unfollowReq.mutate(entryData?.author.pubKey);
-    }
-  };
-
-  const isFollowing = followedProfiles.includes(entryData?.author?.pubKey);
 
   const handleLoadMore = () => {
     if (reqComments.isSuccess && reqComments.hasNextPage && loginState?.fromCache) {
       reqComments.fetchNextPage();
     }
-  };
-
-  const handleMentionClick = (pubKey: string) => {
-    navigateTo?.({
-      appName: '@akashaorg/app-profile',
-      getNavigationUrl: navRoutes => `${navRoutes.rootRoute}/${pubKey}`,
-    });
-  };
-
-  const handleTagClick = (name: string) => {
-    navigateTo?.({
-      appName: '@akashaorg/app-akasha-integration',
-      getNavigationUrl: navRoutes => `${navRoutes.Tags}/${name}`,
-    });
-  };
-
-  const handleAvatarClick = (ev: React.MouseEvent<HTMLDivElement>, pubKey: string) => {
-    ev.preventDefault();
-    navigateTo?.({
-      appName: '@akashaorg/app-profile',
-      getNavigationUrl: routes => `${routes.rootRoute}/${pubKey}`,
-    });
   };
 
   const handleEntryFlag = (entryId: string, itemType: string) => () => {
@@ -146,23 +89,6 @@ const PostPage: React.FC<IPostPageProps & RootComponentProps> = props => {
     }
     props.navigateToModal({ name: 'report-modal', entryId, itemType });
   };
-
-  const handleRepost = (_withComment: boolean, entryId: string) => {
-    analyticsActions.trackEvent({
-      category: AnalyticsCategories.POST,
-      action: 'Repost Clicked',
-    });
-    if (!loginState?.ethAddress) {
-      showLoginModal();
-      return;
-    } else {
-      props.singleSpa.navigateToUrl(
-        `${window.location.origin}/@akashaorg/app-akasha-integration/post/${entryId}?action=repost`,
-      );
-    }
-  };
-
-  const handleEntryNavigate = useEntryNavigation(navigateTo, postId);
 
   const handleFlipCard = () => {
     setShowAnyway(true);
@@ -176,116 +102,8 @@ const PostPage: React.FC<IPostPageProps & RootComponentProps> = props => {
     });
   };
 
-  const handlePostRemove = (commentId: string) => {
-    props.navigateToModal({
-      name: 'entry-remove-confirmation',
-      entryType: EntityTypes.ENTRY,
-      entryId: commentId,
-    });
-  };
-
   const handlePlaceholderClick = () => {
     showLoginModal();
-  };
-
-  const showEditButton = React.useMemo(
-    () => loginState.isReady && loginState.ethAddress === entryData?.author?.ethAddress,
-    [entryData?.author?.ethAddress, loginState.ethAddress, loginState.isReady],
-  );
-
-  const showEntry = React.useMemo(
-    () => !entryData?.delisted && (!isReported || (isReported && entryData?.moderated)),
-    [entryData?.delisted, entryData?.moderated, isReported],
-  );
-
-  const entryUi = () => {
-    if (loginState?.ethAddress) {
-      switch (action) {
-        case 'repost':
-          return (
-            <Extension
-              name="inline-editor_repost"
-              uiEvents={props.uiEvents}
-              data={{ embedEntry: entryData.entryId, isShown: true }}
-            />
-          );
-        case 'edit':
-          return (
-            <Extension
-              name="inline-editor_postedit"
-              uiEvents={props.uiEvents}
-              data={{ entryId: entryData.entryId, action: 'edit', isShown: true }}
-            />
-          );
-      }
-    }
-    return (
-      <Box pad={{ bottom: 'small' }} border={{ side: 'bottom', size: '1px', color: 'border' }}>
-        <EntryBox
-          isRemoved={entryData.isRemoved}
-          entryData={entryData}
-          sharePostLabel={t('Share Post')}
-          shareTextLabel={t('Share this post with your friends')}
-          sharePostUrl={`${window.location.origin}${routes[POST]}/`}
-          onClickAvatar={(ev: React.MouseEvent<HTMLDivElement>) =>
-            handleAvatarClick(ev, entryData.author.pubKey)
-          }
-          repliesLabel={t('Replies')}
-          repostsLabel={t('Reposts')}
-          repostLabel={t('Repost')}
-          repostWithCommentLabel={t('Repost with comment')}
-          shareLabel={t('Share')}
-          copyLinkLabel={t('Copy Link')}
-          flagAsLabel={t('Report Post')}
-          loggedProfileEthAddress={loginState.isReady && loginState?.ethAddress}
-          locale={locale}
-          showMore={true}
-          profileAnchorLink={'/profile'}
-          repliesAnchorLink={routes[POST]}
-          onRepost={handleRepost}
-          onEntryFlag={handleEntryFlag(entryData.entryId, 'post')}
-          handleFollowAuthor={handleFollow}
-          handleUnfollowAuthor={handleUnfollow}
-          isFollowingAuthor={isFollowing}
-          onContentClick={handleEntryNavigate}
-          navigateTo={navigateTo}
-          contentClickable={true}
-          onMentionClick={handleMentionClick}
-          onTagClick={handleTagClick}
-          moderatedContentLabel={t('This content has been moderated')}
-          ctaLabel={t('See it anyway')}
-          handleFlipCard={handleFlipCard}
-          scrollHiddenContent={true}
-          onEntryRemove={handlePostRemove}
-          removeEntryLabel={t('Delete Post')}
-          removedByMeLabel={t('You deleted this post')}
-          removedByAuthorLabel={t('This post was deleted by its author')}
-          disableReposting={entryData.isRemoved}
-          disableReporting={loginState.waitForAuth || loginState.isSigningIn}
-          modalSlotId={props.layoutConfig.modalSlotId}
-          headerMenuExt={
-            showEditButton && (
-              <Extension
-                name={`entry-card-edit-button_${entryData.entryId}`}
-                style={{ width: '100%' }}
-                uiEvents={props.uiEvents}
-                data={{ entryId: entryData.entryId, entryType: EntityTypes.ENTRY }}
-              />
-            )
-          }
-          actionsRightExt={
-            <Extension
-              name={`entry-card-actions-right_${entryData.entryId}`}
-              uiEvents={props.uiEvents}
-              data={{
-                entryId: entryData.entryId,
-                entryType: EntityTypes.ENTRY,
-              }}
-            />
-          }
-        />
-      </Box>
-    );
   };
 
   return (
@@ -319,7 +137,19 @@ const PostPage: React.FC<IPostPageProps & RootComponentProps> = props => {
               handleFlipCard={handleFlipCard}
             />
           )}
-          {showEntry && entryUi()}
+          {
+            <PostEntry
+              postId={postId}
+              loginState={loginState}
+              uiEvents={props.uiEvents}
+              plugins={props.plugins}
+              singleSpa={props.singleSpa}
+              layoutConfig={props.layoutConfig}
+              entryData={entryData}
+              navigateToModal={props.navigateToModal}
+              showLoginModal={props.showLoginModal}
+            />
+          }
 
           {!loginState?.ethAddress && (
             <Box margin="medium">
